operator = is done

+++ CVRP/src/Utillis.cpp
@@ -7,12 +7,12 @@ float Utillis::CalcTourDistance(std::vector<std::vector<vec2>> trucksTour)
 {
 	float totalDistance = 0;
 
-	for (int i = 0; i < trucksTour.size(); i++)
+	for (size_t  i = 0; i < trucksTour.size(); i++)
 	{
 		std::vector<vec2> currentTruckTour = trucksTour[i];
 		int numOfCities = currentTruckTour.size();
 
-		for (int j = 0; j < numOfCities; j++)
+		for (int  j = 0; j < numOfCities; j++)
 		{
 			vec2 start = currentTruckTour[j];
 			vec2 end = currentTruckTour[j + 1 < numOfCities ? j + 1 : 0];
@@ -71,16 +71,16 @@ void Utillis::UpdateSolution(Soulution & mySoulution, std::vector<std::vector<ve
 
 void Utillis::PrintSolution(Soulution & mySoulution)
 {
-	for (int i = 0; i < mySoulution.getTrucksTour().size(); i++)								// printing the path of every truck
+	for (size_t  i = 0; i < mySoulution.getTrucksTour().size(); i++)								// printing the path of every truck
 	{
-		for (int k = 0; k < mySoulution.getTrucksTour()[i].size(); k++)
+		for (size_t  k = 0; k < mySoulution.getTrucksTour()[i].size(); k++)
 		{
 			std::cout << mySoulution.getTrucksTour()[i][k].getIndex() << " " ;
 		}
 		std::cout << mySoulution.getTrucksTour()[0][0].getIndex() << std::endl;
 	}
 
-	for (int i = 0; i < mySoulution.getNumOfCarsAllowed() - mySoulution.getTrucksTour().size(); i++) // printing the path of the rest 
+	for (size_t  i = 0; i < mySoulution.getNumOfCarsAllowed() - mySoulution.getTrucksTour().size(); i++) // printing the path of the rest 
 	{																								 // of the rest of the cars
 		std::cout << mySoulution.getTrucksTour()[0][0].getIndex() << " ";
 		std::cout << mySoulution.getTrucksTour()[0][0].getIndex() << std::endl;
